Update DSA_P055. DEM CAP SO.cpp

+++ BÀI THỰC HÀNH SỐ 2 (Chủ đề Ngăn xếp, Hàng đợi, Duyệt đồ thị và Cây nhị phân) (Thi ngày 07-05-2023)/DSA_P055. DEM CAP SO.cpp
@@ -61,7 +61,7 @@ int main() {
         if (l[i] != -1) res+= 1;
         if (r[i] != n) res+= 1;
         int x = cnt(m[a[i]], first[a[i]], r[i] - 1);
-        res += (x - 1) * x / 2;
+        res += (long long) (x - 1) * x / 2;
         first[a[i]] += x;
     }
     cout << res;
